leds-sysfs: don't use where it lacks timer trigger
Some devices (such as FP2) does not have timer trigger on their LEDs.
Checks for it in usable() probe, so that we don't fail later when
trying to set this attribute later.

For now, failback to leds-hybris backend, as sometimes lights HAL has
an alternative way to blink. It's possible to write a code that controls
LEDs blinking from software, but that means threading work and I'm too
lazy for now.